+++ demos/manual/src/main/java/is/codion/manual/common/demo/ApplicationPanel.java
@@ -25,7 +25,6 @@ import is.codion.common.value.Value;
 import is.codion.plugin.flatlaf.intellij.themes.monokaipro.MonokaiPro;
 import is.codion.swing.common.model.component.combobox.FilterComboBoxModel;
 import is.codion.swing.common.model.component.combobox.FilterComboBoxModel.ItemFinder;
-import is.codion.swing.common.ui.Sizes;
 import is.codion.swing.common.ui.component.Components;
 import is.codion.swing.common.ui.component.combobox.Completion;
 import is.codion.swing.common.ui.component.indicator.ValidIndicatorFactory;
@@ -226,8 +225,8 @@ public final class ApplicationPanel extends JPanel {
 						.build(inputPanel::add);
 
 		FilterComboBoxModel<Item<Integer>> integerItemComboBoxModel = model.createIntegerItemComboBoxModel();
-		Value<Integer> integerItemSelectorValue = integerItemComboBoxModel.createSelectorValue(new IntegerItemFinder());
-		NumberField<Integer> integerItemSelectorField = integerField(integerItemSelectorValue)
+		Value<Integer> integerItemSelector = integerItemComboBoxModel.createSelector(new IntegerItemFinder());
+		NumberField<Integer> integerItemSelectorField = integerField(integerItemSelector)
 						.columns(2)
 						.horizontalAlignment(SwingConstants.CENTER)
 						.selectAllOnFocusGained(true)
@@ -336,8 +335,6 @@ public final class ApplicationPanel extends JPanel {
 										.build(component -> add(component, BorderLayout.SOUTH)))
 						.add(LookAndFeelComboBox.builder().build())
 						.build(southPanel -> add(southPanel, BorderLayout.SOUTH));
-
-		Sizes.setPreferredWidth(this, 400);
 	}
 
 	private static void selectString(ActionEvent event) {